--- UI/_builder/Tmpl/expressions/_private/Helpers.ts
@@ -1,74 +1,0 @@
-/// <amd-module name="UI/_builder/Tmpl/expressions/_private/Helpers" />
-
-import { ProgramNode, IdentifierNode, MemberExpressionNode, Walker } from 'UI/_builder/Tmpl/expressions/_private/Nodes';
-import { IParser } from 'UI/_builder/Tmpl/expressions/_private/Parser';
-
-/**
- * @description Represents Mustache expression helper functions.
- * @author Крылов М.А.
- * @file UI/_builder/Tmpl/expressions/_private/Helpers.ts
- */
-
-export function hasBindings(program: ProgramNode): boolean {
-   if (typeof program.string !== 'string') {
-      return false;
-   }
-   return program.string.indexOf('|mutable') > -1 || program.string.indexOf('|bind') > -1;
-}
-
-export function containsIdentifiers(program: ProgramNode, identifiers: string[], fileName: string): boolean {
-   let hasLocalIdentifier = false;
-   const callbacks = {
-      Identifier: (data: IdentifierNode): void => {
-         if (identifiers.indexOf(data.name) > -1) {
-            hasLocalIdentifier = true;
-         }
-      }
-   };
-   const walker = new Walker(callbacks);
-   program.accept(walker, {
-      fileName
-   });
-   return hasLocalIdentifier;
-}
-
-export function collectIdentifiers(program: ProgramNode, fileName: string): string[] {
-   const identifiers: string[] = [];
-   const callbacks = {
-      Identifier: (node: IdentifierNode): void => {
-         const identifier = node.name;
-         // Do not produce duplicates
-         if (identifiers.indexOf(identifier) === -1) {
-            identifiers.push(node.name);
-         }
-      }
-   };
-   const walker = new Walker(callbacks);
-   program.accept(walker, {
-      fileName
-   });
-   return identifiers;
-}
-
-export function dropBindProgram(program: ProgramNode, parser: IParser, fileName: string): ProgramNode[] {
-   const programs: ProgramNode[] = [];
-   const callbacks = {
-      Identifier: (node: IdentifierNode): void => {
-         programs.push(
-            parser.parse(node.name)
-         );
-      },
-      MemberExpression: (node: MemberExpressionNode): void => {
-         programs.push(
-            parser.parse(node.string)
-         );
-      }
-   };
-   const walker = new Walker(callbacks);
-   program.accept(walker, {
-      fileName
-   });
-   // We need to return value-program and object-program.
-   // Ex. for "a.b.c.d.e" we only return "a.b.c.d" and "a.b.c.d.e".
-   return programs.slice(-2);
-}
